+++ Cargo.toml
@@ -12,4 +12,4 @@ yew = "^0.19.3"
 gloo = "^0.8.0"
 serde = "^1.0.147"
 serde_json = "^1.0.87"
-
+stylist = "^0.10.1"
